- рефакторинг

+++ src/FAST/FAST.h
@@ -5,13 +5,13 @@
 #include <QVector2D>
 #include <QVector>
 
-// мозможно когда-нибудь пригодится
+// возможно когда-нибудь пригодится
 enum class SupportedRadius { three };
 
 int SupportedRadiusToInt(const SupportedRadius &radius);
 
 QVector<QVector2D>
-FAST(const Matrix<qreal> &brightness, const qreal &t,
+FAST(const Matrix<float> &brightness, const float &t,
      const SupportedRadius &supportedRadius = SupportedRadius::three);
 
 #endif // FAST_H
